Improve ErrorCollector in mpf-system-tests
